Handle binary buffer structure validation in own class
