+++ src/validation/SubtreeValidator.ts
@@ -17,15 +17,15 @@ import { ExtendedObjectsValidators } from "./ExtendedObjectsValidators";
 import { MetadataEntityValidator } from "./metadata/MetadataEntityValidator";
 import { PropertyTableValidator } from "./metadata/PropertyTableValidator";
 
-import { BufferObject } from "../structure/BufferObject";
 import { Subtree } from "../structure/Subtree";
-import { BufferView } from "../structure/BufferView";
 import { Availability } from "../structure/Availability";
 import { TileImplicitTiling } from "../structure/TileImplicitTiling";
 
 import { JsonValidationIssues } from "../issues/JsonValidationIssues";
 import { IoValidationIssues } from "../issues/IoValidationIssue";
 import { StructureValidationIssues } from "../issues/StructureValidationIssues";
+import { BinaryBufferStructure } from "./metadata/BinaryBufferStructure";
+import { BinaryBufferStructureValidator } from "./BinaryBufferStructureValidator";
 
 /**
  * A class for validations related to `subtree` objects that have
@@ -409,76 +409,22 @@ export class SubtreeValidator implements Validator<Buffer> {
 
     let result = true;
 
-    // Validate the buffers
-    const buffers = subtree.buffers;
-    const buffersPath = path + "/buffers";
-    if (defined(buffers)) {
-      // The buffers MUST be an array of at least 1 objects
-      if (
-        !BasicValidator.validateArray(
-          buffersPath,
-          "buffers",
-          buffers,
-          1,
-          undefined,
-          "object",
-          context
-        )
-      ) {
-        result = false;
-      } else {
-        // Validate each buffer
-        for (let i = 0; i < buffers!.length; i++) {
-          const buffer = buffers![i];
-          const bufferPath = buffersPath + "/" + i;
-          if (
-            !this.validateBuffer(
-              bufferPath,
-              "buffers/" + i,
-              buffer,
-              hasBinaryBuffer,
-              context
-            )
-          ) {
-            result = false;
-          }
-        }
-      }
-    }
-    // Validate the bufferViews
-    const bufferViews = subtree.bufferViews;
-    const bufferViewsPath = path + "/bufferViews";
-    if (defined(bufferViews)) {
-      //The bufferViews MUST be an array of at least 1 objects
-      if (
-        !BasicValidator.validateArray(
-          bufferViewsPath,
-          "bufferViews",
-          bufferViews,
-          1,
-          undefined,
-          "object",
-          context
-        )
-      ) {
-        result = false;
-      } else {
-        // Validate each bufferView
-        for (let i = 0; i < bufferViews!.length; i++) {
-          const bufferView = bufferViews![i];
-          const bufferViewPath = bufferViewsPath + "/" + i;
-          if (
-            !SubtreeValidator.validateBufferView(
-              bufferViewPath,
-              "bufferViews/" + i,
-              bufferView,
-              context
-            )
-          ) {
-            result = false;
-          }
-        }
-      }
+    // Validate the binary buffer structure, i.e. the `buffers`
+    // and `bufferViews`
+    const binaryBufferStructure: BinaryBufferStructure = {
+      buffers: subtree.buffers,
+      bufferViews: subtree.bufferViews,
+    };
+    const bufferUriIsRequired = !hasBinaryBuffer;
+    if (
+      !BinaryBufferStructureValidator.validateBinaryBufferStructure(
+        path,
+        binaryBufferStructure,
+        bufferUriIsRequired,
+        context
+      )
+    ) {
+      result = false;
     }
 
     // Validate the tileAvailability
@@ -548,203 +494,6 @@ export class SubtreeValidator implements Validator<Buffer> {
       )
     ) {
       result = false;
-    }
-    return result;
-  }
-
-  /**
-   * Performs the validation to ensure that the given object is a
-   * valid `BufferObject` object.
-   *
-   * @param path The path for the `ValidationIssue` instances
-   * @param name The name of the object
-   * @param buffer The `BufferObject` object
-   * @param hasBinaryBuffer Whether the subtree has an internal buffer
-   * (i.e. it was read from a binary subtree file)
-   * @param context The `ValidationContext` that any issues will be added to
-   * @returns Whether the object was valid
-   */
-  private validateBuffer(
-    path: string,
-    name: string,
-    buffer: BufferObject,
-    hasBinaryBuffer: boolean,
-    context: ValidationContext
-  ): boolean {
-    // Make sure that the given value is an object
-    if (!BasicValidator.validateObject(path, name, buffer, context)) {
-      return false;
-    }
-
-    let result = true;
-
-    // Validate the uri
-    const uri = buffer.uri;
-    const uriPath = path + "/uri";
-
-    // When there is no binary buffer (from a binary subtree file),
-    // then the uri MUST be defined
-    if (!hasBinaryBuffer && !defined(uri)) {
-      const message =
-        `The 'uri' property of a buffer is required ` +
-        `when there is no binary buffer`;
-      const issue = StructureValidationIssues.REQUIRED_VALUE_NOT_FOUND(
-        uriPath,
-        message
-      );
-      context.addIssue(issue);
-      result = false;
-    }
-    if (defined(uri)) {
-      // The uri MUST be a string
-      if (!BasicValidator.validateString(uriPath, "uri", uri, context)) {
-        result = false;
-      }
-    }
-
-    // Validate the byteLength
-    // The byteLength MUST be defined
-    // The byteLength MUST be an integer of at least 1
-    const byteLength = buffer.byteLength;
-    const byteLengthPath = path + "/byteLength";
-    if (
-      !BasicValidator.validateIntegerRange(
-        byteLengthPath,
-        "byteLength",
-        byteLength,
-        1,
-        true,
-        undefined,
-        false,
-        context
-      )
-    ) {
-      result = false;
-    }
-
-    // Validate the name
-    const theName = buffer.name;
-    const namePath = path + "/name";
-    if (defined(theName)) {
-      // The name MUST be a string
-      // The name MUST have a length of at least 1
-      if (
-        !BasicValidator.validateStringLength(
-          namePath,
-          "name",
-          theName,
-          1,
-          undefined,
-          context
-        )
-      ) {
-        result = false;
-      }
-    }
-    return result;
-  }
-
-  /**
-   * Performs the validation to ensure that the given object is a
-   * valid `BufferView` object.
-   *
-   * @param path The path for the `ValidationIssue` instances
-   * @param name The name of the object
-   * @param bufferView The `BufferView` object
-   * @param context The `ValidationContext` that any issues will be added to
-   * @returns Whether the object was valid
-   */
-  private static validateBufferView(
-    path: string,
-    name: string,
-    bufferView: BufferView,
-    context: ValidationContext
-  ): boolean {
-    // Make sure that the given value is an object
-    if (!BasicValidator.validateObject(path, name, bufferView, context)) {
-      return false;
-    }
-
-    let result = true;
-
-    // Validate the buffer
-    // The buffer MUST be defined
-    // The buffer MUST be an integer of at least 0
-    const buffer = bufferView.buffer;
-    const bufferPath = path + "/buffer";
-    if (
-      !BasicValidator.validateIntegerRange(
-        bufferPath,
-        "buffer",
-        buffer,
-        0,
-        true,
-        undefined,
-        false,
-        context
-      )
-    ) {
-      result = false;
-    }
-
-    // Validate the byteOffset
-    // The byteOffset MUST be defined
-    // The byteOffset MUST be an integer of at least 0
-    const byteOffset = bufferView.byteOffset;
-    const byteOffsetPath = path + "/byteOffset";
-    if (
-      !BasicValidator.validateIntegerRange(
-        byteOffsetPath,
-        "byteOffset",
-        byteOffset,
-        0,
-        true,
-        undefined,
-        false,
-        context
-      )
-    ) {
-      result = false;
-    }
-
-    // Validate the byteLength
-    // The byteLength MUST be defined
-    // The byteLength MUST be an integer of at least 1
-    const byteLength = bufferView.byteLength;
-    const byteLengthPath = path + "/byteLength";
-    if (
-      !BasicValidator.validateIntegerRange(
-        byteLengthPath,
-        "byteLength",
-        byteLength,
-        1,
-        true,
-        undefined,
-        false,
-        context
-      )
-    ) {
-      result = false;
-    }
-
-    // Validate the name
-    const theName = bufferView.name;
-    const namePath = path + "/name";
-    if (defined(theName)) {
-      // The name MUST be a string
-      // The name MUST have a length of at least 1
-      if (
-        !BasicValidator.validateStringLength(
-          namePath,
-          "name",
-          theName,
-          1,
-          undefined,
-          context
-        )
-      ) {
-        result = false;
-      }
     }
     return result;
   }
